Criação do layout de áreas (docente, aluno, visitante) em templates/views, juntamente com seus respectivos routers em routes/views/areas
