+++ app/public/styles/site.scss
@@ -1,6 +1,7 @@
 // Variáveis de estilo
 // Sobrescrevem o default do bootstrap
 @import "site/variables";
+@import "site/font";
 
 // Bootstrap
 @import "bootstrap/scss/bootstrap";
