+++ app/public/styles/site/_layout.scss
@@ -1,3 +1,5 @@
+
+
 /* Cover section */
 .cover-carousel {
     height: 70vh;
@@ -51,7 +53,7 @@
         }
     }
 
-    
+
 }
 
 #area_visitante:hover{
@@ -128,7 +130,7 @@
     height: 100px;
     padding: 10px;
     cursor: pointer;
-    
+
 }
 
 #noticias-thumb > div:hover{
