+++ app/public/styles/site/_variables.scss
@@ -1,11 +1,10 @@
 // Override Bootstrap variables in this file, e.g.
-$primary: #000
-// $secondary:     $gray-600 !default;
+$primary: #275E87;
+$secondary: #F89700;
 // $success:       $green !default;
 // $info:          $cyan !default;
 // $warning:       $yellow !default;
 // $danger:        $red !default;
 // $light:         $gray-100 !default;
 // $dark:          $gray-800 !default;
-
