+++ app/templates/views/includes/_footer.html
@@ -1,51 +1,40 @@
-<footer class="footer bg-dark text-light mt-0">
-        <!--Footer Links-->
-        <div class="container text-center text-md-left">
-            <div class="row justify-content-center">
+<footer class="footer bg-dark text-light pt-3 mt-5">
+	<!--Footer Links-->
+	<div class="container text-center text-md-left">
+		<div class="row justify-content-center">
 
-                <div class="col-md-6 col-lg-6 col-xl-4 mt-2">
-                    <h6 class="text-uppercase font-weight-bold">
-                        <strong>Departamento de Computação</strong>
-                    </h6>
-                    <hr class="mb-2 mt-0" style="width: 100%; border-color:white;">
-                    <p>Universidade Federal de São Carlos</p>
-                    <p>Rod. Washington Luís, km 235 - Jardim Guanabara</p>
-                    <p>CEP 13565-905 - São Carlos/SP</p>
-                    <p>E-mail: dc@ufscar.br</p>
-                    <p>Telefone: +55 16 3351-8232</p>
-                    <p>Fax: +55 16 3351-8233</p>
-                </div>
+			<div class="col-md-6 col-lg-6 col-xl-4 mt-2">
+				<h6 class="text-uppercase font-weight-bold">
+					<strong>Departamento de Computação - UFSCar</strong>
+				</h6>
 
-                <div class="col-md-6 col-lg-6 col-xl-4 mt-2">
-                    <h6 class="text-uppercase font-weight-bold">
-                        <strong>Outra Coisa</strong>
-                    </h6>
-                    <hr class="mb-2 mt-0" style="width: 100%; border-color:white;">
+        <hr class="mb-2 mt-0" style="width: 100%; border-color:white;">
 
-                </div>
+				<p>Universidade Federal de São Carlos</p>
+				<p>Rod. Washington Luís, km 235 - Jardim Guanabara</p>
+				<p>CEP: 13565-905 - São Carlos/SP</p>
+				<p>E-mail: dc@ufscar.br</p>
+				<p>Telefone: +55 16 3351-8232</p>
+				<p>Fax: +55 16 3351-8233</p>
+			</div>
 
+			<div class="col-md-8 col-offset-1">
+				<iframe
+					height="150px"
+					width="100%"
+					src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3699.8238490302215!2d-47.88273818439348!3d-21.979722710974197!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x94b8711e42fdad05%3A0x6a0cd2ccbce13d28!2sDepartamento+de+Computa%C3%A7%C3%A3o%2C+Universidade+Federal+de+S%C3%A3o+Carlos!5e0!3m2!1sen!2sus!4v1480425559308"
+					frameborder="0"
+					allowfullscreen="allowfullscreen"></iframe>
+			</div>
 
-                <div class="col-md-8 col-lg-6 col-xl-4 mt-2 d-flex justify-content-between align-items-center">
-                    <div>
-                        <img width="150px" src="/images/logos/logo_DC_branco.png" alt="">
-                    </div>
+		</div>
+	</div>
+	<!--/.Footer Links-->
 
-                    <div>
-                        <img width="150px" src="/images/logos/logo_ufscar_branco.png" alt="">
-                    </div>
+	<!--Copyright-->
+	<div class="footer-copyright py-3 text-center">
+		© 2018 Departamento de Computação - UFSCar. Todos os direitos reservados.
+	</div>
+	<!--/.Copyright-->
 
-                </div>
-
-            </div>
-        </div>
-        <!--/.Footer Links-->
-
-
-        <!--Copyright-->
-        <div class="footer-copyright py-3 text-center">
-            © 2018
-            <a href="http://petbcc.dc.ufscar.br"> PET-BCC UFSCar</a>. Todos os direitos reservados.
-        </div>
-        <!--/.Copyright-->
-
-    </footer>
+</footer>
