+++ app/templates/views/includes/_navbar.html
@@ -1,4 +1,4 @@
-<nav class="navbar fixed-top navbar-expand-lg navbar-dark bg-primary"> 
+<nav class="navbar  navbar-expand-lg navbar-dark bg-primary"> 
 
     <a class="navbar-brand" href="/">
         <img src="/images/logos/logo_DC_branco.png" width="60" alt="Logo do departamento de computação">
