+++ app/templates/views/index.html
@@ -236,7 +236,17 @@
     </div>
 
 </main>
+{% endblock %}
 
-<iframe class="mb-0 mt-5" height="300px" width="100%" src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d3699.8238490302215!2d-47.88273818439348!3d-21.979722710974197!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x94b8711e42fdad05%3A0x6a0cd2ccbce13d28!2sDepartamento+de+Computa%C3%A7%C3%A3o%2C+Universidade+Federal+de+S%C3%A3o+Carlos!5e0!3m2!1sen!2sus!4v1480425559308" 
-    frameborder="0" allowfullscreen></iframe>
+{% block js %}
+<script>
+  $(document).ready(function () {
+    var noticias_carousel = $('#noticias-carousel');
+    var thumbnails = document.getElementById("noticias-thumb").children;
+    noticias_carousel.on('slide.bs.carousel', function (data) {
+      thumbnails[data.from].classList.remove('active');
+      thumbnails[data.to].classList.add('active');
+    });
+  })
+</script>
 {% endblock %}
