+++ app/templates/views/layouts/default.html
@@ -3,38 +3,37 @@
 <!doctype html>
 <html>
 
-<head>
-	<meta charset="utf-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<meta http-equiv="X-UA-Compatible" content="IE=edge">
+	<head>
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, initial-scale=1.0">
+		<meta http-equiv="X-UA-Compatible" content="IE=edge">
 
-	<title>{{ title|default("DC - UFSCar") }}</title>
-	<link rel="shortcut icon" href="/favicon.ico" type="image/x-icon">
+		<title>{{ title|default("DC - UFSCar") }}</title>
+		<link rel="shortcut icon" href="/favicon.ico" type="image/x-icon">
 
-	<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.9/css/all.css" integrity="sha384-5SOiIsAziJl6AWe0HWRKTXlfcSHKmYV4RBF18PPJ173Kzn7jzMyFuTtk8JA7QQG1"
-	 crossorigin="anonymous">
+		<link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.9/css/all.css" integrity="sha384-5SOiIsAziJl6AWe0HWRKTXlfcSHKmYV4RBF18PPJ173Kzn7jzMyFuTtk8JA7QQG1" crossorigin="anonymous">
 
-	<link href="/styles/site.css" rel="stylesheet"> {# This file provides the default styling for the KeystoneJS Content Editor #}
-	{% if user and user.canAccessKeystone %}
-	<link href="/keystone/styles/content/editor.min.css" rel="stylesheet"> {% endif %}
+		<link href="/styles/site.css" rel="stylesheet">
 
-	<!--[if lt IE 9]>
-			<script src="//cdn.jsdelivr.net/html5shiv/3.7.3/html5shiv.js"></script>
-			<script src="//cdn.jsdelivr.net/respond/1.4.2/respond.min.js"></script>
-	<![endif]-->
+		{# This file provides the default styling for the KeystoneJS Content Editor #}
+		{% if user and user.canAccessKeystone %}
+			<link href="/keystone/styles/content/editor.min.css" rel="stylesheet">
+		{% endif %}
 
-	{# Include template-specific stylesheets by extending the css block #} 
-	{% block css %}{% endblock %} {# Add any other template-specific
-	HEAD tags by extending the head block #} 
-	{% block head %}{% endblock %}
-</head>
+		<!--[if lt IE 9]> <script src="//cdn.jsdelivr.net/html5shiv/3.7.3/html5shiv.js"></script> <script src="//cdn.jsdelivr.net/respond/1.4.2/respond.min.js"></script> <![endif]-->
 
-<body>
-	{# NAVBAR #} 
-	{% include "includes/_navbar.html" %}
-	
-	{# HEADER #}
-	{# <div id="header" class="container"> #}
+		{# Include template-specific stylesheets by extending the css block #}
+		{% block css %}{% endblock %}
+		{# Add any other template-specific HEAD tags by extending the head block #}
+		{% block head %}{% endblock %}
+	</head>
+
+	<body>
+		{# NAVBAR #}
+		{% include "includes/_navbar.html" %}
+
+		{# HEADER #}
+		{# <div id="header" class="container"> #}
 
 		{# Customise your site"s navigation by changing the navLinks Array in ./routes/middleware.js ... or completely change this
 		header to suit your design. #}
@@ -75,45 +74,40 @@
 				</div>
 			</div>
 		</div> #}
-	{# </div> #}
+		{# </div> #}
 
-	{# BODY #}
-	<div id="body">
-		{# NOTE: There is no .container wrapping class around body blocks to allow more flexibility in design. Remember to include
-		it in your templates when you override the intro and content blocks! #} {# The Intro block appears above flash messages
-		(used for temporary information display) #} 
-		{% block intro %}{% endblock %} {# Flash messages allow you to display once-off
-		status messages to users, e.g. form validation errors, success messages, etc. #} {{ FM.flashMessages(messages) }} {# The
-		content block should contain the body of your template"s content #}
-		
-		{% block content %}{% endblock %}
-	</div>
+		{# BODY #}
+		<div id="body">
+			{# NOTE: There is no .container wrapping class around body blocks to allow more flexibility in design. Remember to include
+		it in your templates when you override the intro and content blocks! #}
+			{# The Intro block appears above flash messages
+		(used for temporary information display) #}
 
-	{# FOOTER #}
-	{% include "includes/_footer.html" %}
+			{% block intro %}{% endblock %}
+			{# Flash messages allow you to display once-off
+		status messages to users, e.g. form validation errors, success messages, etc. #}
 
-	{# The KeystoneJS Content Editor provides support for ks-editable data attributes, which generate links to edit content
-	for users who can access Keystone #} 
-	{% if user and user.canAccessKeystone %}
-	<script src="/keystone/js/content/editor.js"></script> 
-	{% endif %}
+			{{ FM.flashMessages(messages) }}
+			{# The content block should contain the body of your template"s content #}
 
-	{# Add scripts that are globally required by your site here. #} {# Include template-specific javascript files
-	by extending the js block #} 
+			{% block content %}{% endblock %}
+		</div>
 
-</body>
-<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
-<script src="/styles/bootstrap/dist/js/bootstrap.min.js"></script>
+		{# FOOTER #}
+		{% include "includes/_footer.html" %}
 
-<script>
-	$(document).ready(function () {
-		var noticias_carousel = $('#noticias-carousel');
-		var thumbnails = document.getElementById("noticias-thumb").children;
-		noticias_carousel.on('slide.bs.carousel', function (data) {
-			thumbnails[data.from].classList.remove('active');
-			thumbnails[data.to].classList.add('active');
-		});
-	})
-</script>
-{% block js %}{% endblock %} 
+		{# The KeystoneJS Content Editor provides support for ks-editable data attributes, which generate links to edit content
+	for users who can access Keystone #}
+		{% if user and user.canAccessKeystone %}
+			<script src="/keystone/js/content/editor.js"></script>
+		{% endif %}
+
+		{# Add scripts that are globally required by your site here. #}
+		{# Include template-specific javascript files
+	by extending the js block #}
+
+	</body>
+	<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"></script>
+	<script src="/bootstrap.min.js"></script>
+	{% block js %}{% endblock %}
 </html>
